+++ lib/binance_client/models/network.rb
@@ -1,27 +1,25 @@
 module BinanceClient
   class Network < BaseModel
 
-    METHODS = %i[
+    PLAIN_METHODS = %i[
       address_regex
       coin
       deposit_desc
-      deposit_enable
-      is_default
       memo_regex
       min_confirm
       name
       network
-      reset_address_status
       special_tips
       un_lock_confirm
       withdraw_desc
-      withdraw_enable
+    ].freeze
+
+    DECIMAL_METHODS = %i[
       withdraw_fee
       withdraw_integer_multiple
       withdraw_max
       withdraw_min
-      same_address
-    ]
+    ].freeze
 
     BOOL_MAP = {
       deposit_enable: :deposit_enabled?,
@@ -30,6 +28,8 @@ module BinanceClient
       withdraw_enable: :withdraw_enabled?,
       same_address: :same_address?
     }.freeze
+
+    METHODS = (PLAIN_METHODS + DECIMAL_METHODS + BOOL_MAP.keys).freeze
 
     attr_accessor :raw_hash
     attr_writer *METHODS
@@ -40,7 +40,7 @@ module BinanceClient
       end
     end
 
-    METHODS.each do |method_name|
+    PLAIN_METHODS.each do |method_name|
       define_method method_name do
         memoize_results "@#{method_name}" do
           raw_hash[method_name.to_s.camelcase(:lower)]
@@ -48,7 +48,21 @@ module BinanceClient
       end
     end
 
+    DECIMAL_METHODS.each do |method_name|
+      define_method method_name do
+        memoize_results "@#{method_name}" do
+          raw_hash[method_name.to_s.camelcase(:lower)].to_d
+        end
+      end
+    end
+
     BOOL_MAP.each do |original_method_name, alias_method_name|
+      define_method original_method_name do
+        memoize_results "@#{original_method_name}" do
+          raw_hash[original_method_name.to_s.camelcase(:lower)]
+        end
+      end
+
       alias_method alias_method_name, original_method_name
     end